licenses, move stuff around
