+++ platform/jvm/capture-plugin/src/main/kotlin/io/bitdrift/capture/AndroidComponentsConfig.kt
@@ -5,6 +5,32 @@
 // LICENSE file or at:
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
+/**
+ * Adapted from https://github.com/getsentry/sentry-android-gradle-plugin/tree/4.14.1
+ *
+ * MIT License
+ *
+ * Copyright (c) 2020 Sentry
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 package io.bitdrift.capture
 
 import com.android.build.api.instrumentation.AsmClassVisitorFactory
@@ -14,15 +40,17 @@ import com.android.build.api.instrumentation.InstrumentationScope
 import com.android.build.api.variant.AndroidComponentsExtension
 import com.android.build.api.variant.Variant
 import io.bitdrift.capture.CapturePlugin.Companion.sep
+import io.bitdrift.capture.extension.BitdriftPluginExtension
+import io.bitdrift.capture.instrumentation.SpanAddingClassVisitorFactory
 import org.gradle.api.Project
 import java.io.File
 
 fun AndroidComponentsExtension<*, *, *>.configure(
-        project: Project,
-        extension: BitdriftPluginExtension,
+    project: Project,
+    extension: BitdriftPluginExtension,
 ) {
-    // temp folder for sentry-related stuff
-    val tmpDir = File("${project.buildDir}${sep}tmp${sep}sentry")
+    // Temp folder for outputting debug logs
+    val tmpDir = File("${project.layout.buildDirectory}${sep}tmp${sep}bitdrift")
     tmpDir.mkdirs()
 
     onVariants { variant ->
